Update 1.1
Models, views and templates were rewritten and, when appropriate,
relocated to their designated folders. This new structure reflects the
new underlying data aggregation with all the different apps.

Major change in how albums are stored. Now an album is connected with a
“faixa” class, which in turn has all title, number, and file as its
properties.

A change in the JavaScript for the player now ensures that when a track
is done playing the icon displayed is for replay.

Change in the search feature to allow search across all models with the
results being displayed in a well structure manner.

Old database data has been flushed away. Only new and accurate content
from now on. No more ‘Lorem Ipsum’, in other words.

— Leo N.